+++ packages/iov-bns/src/types.ts
@@ -207,16 +207,6 @@ export interface AddAddressToUsernameTx extends UnsignedTransaction {
   readonly payload: ChainAddressPair;
 }
 
-/**
- * Associates a simple name to an account on a weave-based blockchain.
- *
- * @deprecated will be dropped in favour of RegisterUsernameTx
- */
-export interface SetNameTx extends UnsignedTransaction {
-  readonly kind: "bns/set_name";
-  readonly name: string;
-}
-
 export interface RegisterBlockchainTx extends UnsignedTransaction {
   readonly kind: "bns/register_blockchain";
   /**
@@ -259,7 +249,6 @@ export type BnsTx =
   | SwapTimeoutTransaction
   // BNS
   | AddAddressToUsernameTx
-  | SetNameTx
   | RegisterBlockchainTx
   | RegisterUsernameTx
   | RemoveAddressFromUsernameTx;
@@ -282,10 +271,6 @@ export function isAddAddressToUsernameTx(
   transaction: UnsignedTransaction,
 ): transaction is AddAddressToUsernameTx {
   return isBnsTx(transaction) && transaction.kind === "bns/add_address_to_username";
-}
-
-export function isSetNameTx(transaction: UnsignedTransaction): transaction is SetNameTx {
-  return isBnsTx(transaction) && transaction.kind === "bns/set_name";
 }
 
 export function isRegisterBlockchainTx(