+++ packages/iov-bns/types/index.d.ts
@@ -2,4 +2,4 @@ export { bnsCodec } from "./bnscodec";
 export { bnsConnector } from "./bnsconnector";
 export { BnsConnection } from "./bnsconnection";
 export { bnsNonceTag, bnsSwapQueryTags } from "./tags";
-export { BnsBlockchainNft, BnsBlockchainsByChainIdQuery, BnsBlockchainsQuery, BnsUsernamesByChainAndAddressQuery, BnsUsernamesByOwnerAddressQuery, BnsUsernamesByUsernameQuery, BnsUsernamesQuery, BnsUsernameNft, BnsTx, AddAddressToUsernameTx, SetNameTx, RegisterBlockchainTx, RegisterUsernameTx, RemoveAddressFromUsernameTx, } from "./types";
+export { BnsBlockchainNft, BnsBlockchainsByChainIdQuery, BnsBlockchainsQuery, BnsUsernamesByChainAndAddressQuery, BnsUsernamesByOwnerAddressQuery, BnsUsernamesByUsernameQuery, BnsUsernamesQuery, BnsUsernameNft, BnsTx, AddAddressToUsernameTx, RegisterBlockchainTx, RegisterUsernameTx, RemoveAddressFromUsernameTx, } from "./types";