+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.12.0
+
+Breaking changes
+
+* @iov/bcp-type: Remove `Bcp*` prefix from `BcpAccount`, `BcpAccountQuery`,
+  `BcpAddressQuery`, `BcpPubkeyQuery`
+* @iov/bcp-type: Remove `Account.name` in favour of BNS' username NFTs.
+* @iov/bns: Remove obsolete `SetNameTx`
+
 ## 0.11.0
 
 * @iov/bcp-type: `BcpConnection.getNonce` now returns a `Promise<Nonce>` and